Add version autodetection to character sheets
The game version should be autodetected from normal character codes
now and the required branch printed at the bottom of the sheet.

Hand-modified codes, or codes where the player did not spend all their
attribute points, will not have a required branch printed and will
be assumed to be from beta and up.