Move wiki title search to different API
With the migration of the Caves of Qud wiki to the Unified Community
Platform on Gamepedia, title search via `intitle:` is no longer
possible. While researching this, I found that the 'search' API path can
be replaced with 'opensearch' and skip the step of retrieving URLs for
the results. This new ?wiki command returns the same results as it did
after the UCP migration, but with less complexity. The results are also
better than *before* the UCP migration either way, since fuzzy search
is now the default for 'search' and is used here with 'opensearch'.